+++ Mods/S7_Journal_6b1aee95-bd4c-4190-a919-4aa7096eaa2b/Story/RawFiles/Lua/DebugMode.lua
@@ -11,10 +11,15 @@ local function listPersistentJournals()
     if PersistentVars.JournalData then for fileName, content in pairs(PersistentVars.JournalData) do S7DebugPrint(fileName, "DebugMode") end end
 end
 
-Ext.RegisterConsoleCommand("S7_Journal", function (cmd, command)
+Ext.RegisterConsoleCommand("S7_Journal", function (cmd, command, ...)
+    local args = {...}
     if command == "AddJournal" then addJournal() end
     if command == "ResyncSettings" then ResynchronizeModSettings() end
     if command == "ListPersistentJournals" then listPersistentJournals() end
+    if command == "RemoveJournalData" then
+        if string.lower(tostring(args[1])) == "all" then PersistentVars.JournalData = nil
+        elseif PersistentVars.JournalData[tostring(args[1])] ~= nil then PersistentVars.JournalData[tostring(args[1])] = nil end
+    end
 end)
 
 -- Ext.RegisterOsirisListener("GameStarted", 2, "after", function (level, isEditorMode)